Move logic to seed databasefrom the server index.js database file to it's own file
