+++ .eslintrc.js
@@ -6,5 +6,5 @@
  */
 
 module.exports = {
-  extends: './node_modules/eslint-config-hackreactor/index.js'
+  extends:  './node_modules/eslint-config-hackreactor/index.js'
 };
