+++ package.json
@@ -10,7 +10,7 @@
   "scripts": {
     "react-dev": "webpack -d --watch --display-error-details",
     "start": "nodemon server/index.js",
-    "seed-db": "node database/db.js"
+    "seed-db": "node database/seeder.js"
   },
   "dependencies": {
     "@babel/cli": "^7.2.3",
@@ -28,7 +28,10 @@
     "@babel/preset-env": "^7.3.1",
     "@babel/preset-react": "^7.0.0",
     "babel-loader": "^8.0.5",
+    "eslint": "^5.13.0",
+    "eslint-config-airbnb-base": "^13.1.0",
     "eslint-config-hackreactor": "git://github.com/reactorcore/eslint-config-hackreactor.git",
+    "eslint-plugin-import": "^2.16.0",
     "faker": "^4.1.0",
     "nodemon": "^1.18.9",
     "webpack": "^4.29.3",
