+++ server/controller.js
@@ -9,5 +9,5 @@ module.exports = {
       }
       res.send(data);
     });
-  }
+  },
 };
